feature_engineering/extraction/requirements.txt
Flask_RESTful==0.3.8
grpcio==1.31.0
tensorflow_serving_api==2.3.0
numpy==1.18.5
Flask==1.1.2
tensorflow==2.3.1
